backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
openai
pypdf
requests
